requirements.txt
ansible==2.3.0.0
appdirs==1.4.3
Jinja2==2.9.6
nose==1.3.7
packaging==16.8
parameterized==0.6.1
pika==0.10.0
pyparsing==2.2.0
six==1.10.0
virtualenv==15.1.0
funcsigs==1.0.2
mock==2.0.0
pbr==3.0.1
six==1.10.0
psycopg2==2.7.1
python-dateutil==2.6.0
